pyproject.toml
[tool.poetry]
name = "jeeves"
version = "0.1.0"
description = "Jeeves is my personal assistant."
authors = ["Joe Banks <joe@jb3.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.19.2"
loguru = "^0.6.0"
"discord.py" = {url = "https://github.com/Rapptz/discord.py/archive/6dc314628ea4e349c8fdf8921ae5f080410dc2fb.zip"}
taskipy = "^1.10.1"
icmplib = "^3.0.3"
tabulate = "^0.8.9"
ansi = "^0.3.6"

[tool.poetry.dev-dependencies]

[tool.taskipy.tasks]
start = "python -m jeeves"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
